Adjustment to add an extension class that has particular methods called during certain execution points of a query.
The class QuerierExtension has been added with the intent that any systems which wish to interact with the queries will subclass it and pass their extension to the WebServiceQuerier construction call or when executing the send method of subclasses of WebServiceQuery. To support some desired logging this also resulted in the creation of some read accessor methods and adjustment to the visibility of a few properties in WebServiceQuery and WebServiceQuerier.